requirements/base.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.6
aiohttp-proxy==0.1.2
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
asgiref==3.7.2
astroid==2.15.5
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
billiard==4.1.0
celery==5.3.1
celery-progress==0.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dill==0.3.6
Django==4.2.5
django-celery-results==2.5.1
django-cors-headers==3.13.0
django-debug-toolbar==4.0.0
django-jinja==2.10.2
django-phonenumber-field==7.1.0
django-redis==5.3.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
flake8==6.0.0
flower==2.0.1
frozenlist==1.4.0
humanize==4.8.0
idna==3.4
isort==5.12.0
Jinja2==3.1.2
kombu==5.3.1
lazy-object-proxy==1.9.0
loguru==0.7.2
magic-filter==1.0.12
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
openai==0.27.2
openpyxl==3.1.2
peewee==3.17.0
phonenumberslite==8.13.13
Pillow==9.5.0
platformdirs==3.5.3
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.6
psycopg2-binary==2.9.9
pycodestyle==2.10.0
pydantic==2.3.0
pydantic_core==2.6.3
pyflakes==3.0.1
pylint==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.1.0
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
yarl==1.9.2
